+++ src/test/customMatchers.ts
@@ -1,15 +1,6 @@
 import { ContractTransaction } from "ethers";
 import { getVmError } from "../core/contracts/contract";
 
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace jest {
-    interface Matchers<R> {
-      transactionRejectsWith(message: string | RegExp): R;
-    }
-  }
-}
-
 expect.extend({
   async transactionRejectsWith(pendingTx: Promise<ContractTransaction>, message: string | RegExp) {
     try {